Improved the speed and memory usage of midi parsing, made it possible to parse larger files

+++ src/midi.h
@@ -48,7 +48,7 @@ struct midistats {
   __uint8_t max_key;
 };
 
-struct midistats parse_tracks(const struct track_chunk* tracks, const int num_tracks,
+struct midistats parse_tracks(const struct track_chunk* tracks, const short num_tracks,
                               struct instr** instructions, OBJECT* background);
 
 //--------------------------------------------------------------------------------------------------
